+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Not released:
 
 * Add `rotationSnapTolerance` property to `Konva.Transformer`.
+* Add `getActiveAnchor()` method to `Konva.Transformer`
 
 ## 4.1.6 - 2020-02-25